feat: separate status and transition editors
